requirements.txt
Django==1.5.5
Fabric==1.8.0
South==0.8.2
django-tastypie==0.10.0
ecdsa==0.10
paramiko==1.12.0
pycrypto==2.6.1
python-dateutil==2.2
python-mimeparse==0.1.4
six==1.4.1
wsgiref==0.1.2
